refactor: move SVUWithOptions to mage.go again
