+++ catalog/svu/dagger.go
@@ -5,8 +5,6 @@ import (
 	"fmt"
 	"strings"
 
-	"github.com/magefile/mage/mg"
-
 	"github.com/mesosphere/daggers/daggers"
 	"github.com/mesosphere/daggers/daggers/containers"
 )
@@ -69,30 +67,6 @@ func Run(
 		Version:              strings.TrimSpace(version),
 		VersionWithoutPrefix: strings.TrimSpace(versionWithoutPrefix),
 	}, nil
-}
-
-// SVUWithOptions runs svu with specific options.
-//
-// TODO: Refactor this to make it more generic and reusable. Temporary solution to get svu working.
-//
-//nolint:revive // Stuttering is fine here to provide a functional options variant of SVU call.
-func SVUWithOptions(ctx context.Context, opts ...daggers.Option[config]) error {
-	verbose := mg.Verbose() || mg.Debug()
-
-	runtime, err := daggers.NewRuntime(ctx, daggers.WithVerbose(verbose))
-	if err != nil {
-		return err
-	}
-	defer runtime.Close()
-
-	output, err := Run(ctx, runtime, opts...)
-	if err != nil {
-		return err
-	}
-
-	fmt.Println(output.Version)
-
-	return nil
 }
 
 func flagsFromConfig(cfg *config) []string {
