+++ catalog/svu/mage.go
@@ -2,6 +2,11 @@ package svu
 
 import (
 	"context"
+	"fmt"
+
+	"github.com/magefile/mage/mg"
+
+	"github.com/mesosphere/daggers/daggers"
 )
 
 // Current runs svu current.
@@ -27,4 +32,26 @@ func Minor(ctx context.Context) error {
 // Patch runs svu patch.
 func Patch(ctx context.Context) error {
 	return SVUWithOptions(ctx, WithCommand(CommandPatch))
+}
+
+// SVUWithOptions runs svu with specific options.
+//
+//nolint:revive // Stuttering is fine here to provide a functional options variant of SVU call.
+func SVUWithOptions(ctx context.Context, opts ...daggers.Option[config]) error {
+	verbose := mg.Verbose() || mg.Debug()
+
+	runtime, err := daggers.NewRuntime(ctx, daggers.WithVerbose(verbose))
+	if err != nil {
+		return err
+	}
+	defer runtime.Close()
+
+	output, err := Run(ctx, runtime, opts...)
+	if err != nil {
+		return err
+	}
+
+	fmt.Println(output.Version)
+
+	return nil
 }
